+++ Sources/Alchemy/Rune/Relationships/Types/Relationship.swift
@@ -1,0 +1,7 @@
+import NIO
+
+public protocol Relationship {
+    associatedtype From: Model
+    associatedtype To: RelationAllowed
+    func load(_ from: [From], from eagerLoadKeyPath: KeyPath<From, Self>) -> EventLoopFuture<[From]>
+}
